restructuration for easier maintenability: separation of the style, creation of src folder
